stockproj/requirements.txt
vnstock>=3.2.6
plotly>=6.3.0
bokeh==2.4.3
streamlit-bokeh-events>=0.1.2
pandas>=2.3.1
numpy
python-dateutil>=2.9.0.post0
### stockmarket-pattern-detection-yolov8
mss==10.0.0 
opencv-python==4.11.0.86
ultralytics==8.3.94
openpyxl==3.1.5
kaleido >=1.0.0